Refactor dir name
