+++ src/components/Tools/Form.tsx
@@ -30,7 +30,7 @@ import { useTranslation } from 'react-i18next'
 import { useNavigate } from 'react-router-dom'
 import { Avatar } from '~components/Images/Avatar'
 import { Image, ServerImage } from '~components/Images/ServerImage'
-import { useDefaultUserCommunities } from '~components/communities/queries'
+import { useDefaultUserCommunities } from '~components/Communities/queries'
 import { useInfoContext } from '~components/Layout/Contexts/InfoContext'
 import FormSubmitMessage from '~components/Layout/Form/FormSubmitMessage'
 import { LocationPicker } from '~components/Layout/Map/LocationPicker'
