+++ src/components/Users/UserCommunities.tsx
@@ -5,35 +5,29 @@ import { Box, Heading, Stack, Text, useColorModeValue } from '@chakra-ui/react'
 import { LoadingSpinner } from '~components/Layout/LoadingSpinner'
 import { ElementNotFound } from '~components/Layout/ElementNotFound'
 import { icons } from '~theme/icons'
-import { useUserCommunities } from '~components/communities/queries'
+import { useUserCommunities } from '~components/Communities/queries'
 import { useAuth } from '~components/Auth/AuthContext'
-import { CommunityCard } from '~components/communities/Card'
+import { CommunityCard } from '~components/Communities/Card'
 
 export const UserCommunities = () => {
   const { t } = useTranslation()
   const { id } = useParams<{ id: string }>()
   const { user: currentUser } = useAuth()
   const isCurrentUser = id === currentUser?.id
-  
+
   const { data: communities, isLoading, isError } = useUserCommunities(id!)
-  
+
   const bgColor = useColorModeValue('white', 'gray.800')
   const borderColor = useColorModeValue('gray.200', 'gray.700')
-  
+
   if (isLoading) {
     return <LoadingSpinner />
   }
-  
+
   if (isError) {
-    return (
-      <ElementNotFound
-        icon={icons.users}
-        title={t('common.error')}
-        desc={t('common.something_went_wrong')}
-      />
-    )
+    return <ElementNotFound icon={icons.users} title={t('common.error')} desc={t('common.something_went_wrong')} />
   }
-  
+
   if (!communities || communities.length === 0) {
     return (
       <ElementNotFound
@@ -43,11 +37,11 @@ export const UserCommunities = () => {
       />
     )
   }
-  
+
   return (
     <Stack spacing={6}>
       <Stack spacing={4}>
-        {communities.map(community => (
+        {communities.map((community) => (
           <CommunityCard key={community.id} community={community} />
         ))}
       </Stack>
