+++ src/pages/communities/detail.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect } from 'react'
 import { useOutletContext, useParams } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
-import { CommunityDetail } from '~components/communities/Detail'
+import { CommunityDetail } from '~components/Communities/Detail'
 import { TitlePageLayoutContext } from '~src/pages/TitlePageLayout'
-import { useCommunityDetail } from '~components/communities/queries'
+import { useCommunityDetail } from '~components/Communities/queries'
 import { LoadingSpinner } from '~components/Layout/LoadingSpinner'
 import { ElementNotFound } from '~components/Layout/ElementNotFound'
 import { icons } from '~theme/icons'
