+++ src/pages/communities/edit.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect } from 'react'
 import { useOutletContext, useParams } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
-import { CommunityForm } from '~components/communities/Form'
+import { CommunityForm } from '~components/Communities/Form'
 import { FormLayoutContext } from '~src/pages/FormLayout'
-import { useCommunityDetail } from '~components/communities/queries'
+import { useCommunityDetail } from '~components/Communities/queries'
 import { LoadingSpinner } from '~components/Layout/LoadingSpinner'
 import { ElementNotFound } from '~components/Layout/ElementNotFound'
 import { icons } from '~theme/icons'
@@ -13,7 +13,7 @@ export const Edit = () => {
   const { id } = useParams<{ id: string }>()
   const { setTitle } = useOutletContext<FormLayoutContext>()
   const { data: community, isLoading, isError } = useCommunityDetail(id!)
-  
+
   useEffect(() => {
     // Set the page title when community data is loaded
     if (community) {
@@ -22,20 +22,16 @@ export const Edit = () => {
       setTitle(t('communities.edit_community'))
     }
   }, [community, setTitle, t])
-  
+
   if (isLoading) {
     return <LoadingSpinner />
   }
-  
+
   if (isError || !community) {
     return (
-      <ElementNotFound
-        icon={icons.users}
-        title={t('communities.not_found')}
-        desc={t('communities.not_found_desc')}
-      />
+      <ElementNotFound icon={icons.users} title={t('communities.not_found')} desc={t('communities.not_found_desc')} />
     )
   }
-  
+
   return <CommunityForm initialData={community} isEdit />
 }
