+++ src/pages/communities/view.tsx
@@ -2,13 +2,13 @@ import React, { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
 import { TitlePageLayoutContext } from '~src/pages/TitlePageLayout'
 import { Link as RouterLink, useOutletContext } from 'react-router-dom'
-import { useCommunityInvites, useDefaultUserCommunities, useUserCommunities } from '~components/communities/queries'
+import { useCommunityInvites, useDefaultUserCommunities, useUserCommunities } from '~components/Communities/queries'
 import { RoutedTabs, TabConfig } from '~components/Layout/RoutedTabs'
 import { ROUTES } from '~src/router/routes'
 import { usePendingActions } from '~components/Layout/Contexts/PendingActionsProvider'
 import { BadgeCounter } from '~components/Layout/BadgeIcon'
-import { CommunitiesList } from '~components/communities/List'
-import { CommunityInvites } from '~components/communities/Invites'
+import { CommunitiesList } from '~components/Communities/List'
+import { CommunityInvites } from '~components/Communities/Invites'
 import { Button } from '@chakra-ui/react'
 import { FiPlus } from 'react-icons/fi'
